fixing remote restart
fixing remote restart

fixing remote restart

better piface handling

adding rapsberry pi setup

virtualenv setup

fixing raspberrypi setup

fix raspberry pi setup

fixing restart when called directly

fixing restart when called directly

fixing restart when called directly

Add back raspberry specific stuff

some fixes

fix raspberry pi

some comments

getting to fix the pi

adding exec bit

fixing redirection

fixing redirection

fixing redirection

fixing redirection

fixing redirection

fixing redirection

fixing redirection

fixing redirection

fixing redirection

fixing redirection

fixing redirection

fixing redirection

add rpi button

add rpi button

add rpi button

add rpi button

add rpi button

add rpi button

add rpi button

add rpi button

add rpi button

add rpi button

add rpi button

add rpi button

add rpi button

add rpi button

add rpi button

add rpi button

+++ restartServer.sh
@@ -1,17 +1,29 @@
-#!/bin/sh
+#!/bin/bash
 
-cd $1
+# Disable DPMS / Screen blanking
+xset -dpms
+xset s off
+
+
+# Pass the cwd of the project
+
+if [ $# -gt 0 ]; then
+  cd $1
+fi
 
 serverPid=
 if [ -f server.pid ]; then
     serverPid=$(cat server.pid)
-    echo "Killing $serverPid"
-    ps -ef | grep $serverPid
-    kill -- -$( ps opgid= $serverPid | tr -d ' ' ) || true
     rm server.pid
+    if [ -n $serverPid ]; then
+      echo "Killing $serverPid"
+      if [ $(ps -ef | grep $serverPid | wc -l) -gt 1 ]; then
+        kill -- -$( ps opgid= $serverPid | tr -d ' ' ) || true
+      fi
+    fi
 fi
 
-nohup $HOME/.virtualenvs/pump-quizz/bin/python2.7 server.py > /tmp/pump-quizz.log 2>&1&
+nohup $HOME/.virtualenvs/pump-quizz/bin/python2.7 -u server.py > /tmp/pump-quizz.log 2>&1 &
 
 serverPid=$!
 ps -ef | grep $serverPid
